Encapsulate Process struct
